\name{getbounds-methods}
\alias{getbounds}
\docType{methods}
\alias{getbounds-methods}
\alias{getbounds,sampalgres-method}
\alias{getbounds,sampalgontheflyres-method}
\title{Methods for Function getbounds in Package `simctest'}
\description{
 returns bounds on the p.value if the algorithm has not stopped yet.
}
\usage{
getbounds(data)
}
\arguments{
  \item{data}{an object of type \code{\linkS4class{sampalgres}} or \code{linkS4class{sampalgontheflyres}}.}
}
\keyword{methods}
