\name{sampalg-class}
\docType{class}
\alias{sampalg-class}

\title{Class "sampalg"}
\description{Virtual base class for several sequential sampling algorithms.}
\section{Objects from the Class}{
This is a virtual class - no objects should be derived from it.
}
\section{Slots}{
	 \describe{
    \item{\code{internal}:}{Internal status data of the algorithm. Object of class \code{"environment"}  }
  }
}
\section{Methods}{
No methods defined with class "sampalg" in the signature.
}
\author{ Axel Gandy }

\seealso{
 \code{\linkS4class{sampalgonthefly}}, \code{\linkS4class{sampalgPrecomp}}
}
\keyword{classes}
