\name{mmctSamplerGeneric-class}
\docType{class}
\alias{mmctSamplerGeneric-class}

\title{Class "mmctSamplerGeneric"}
\description{Generic class, has to be implemented as "mmctSampler".}
\section{Objects from the Class}{
This is a virtual class - no objects should be derived from it.
}
\section{Methods}{
  \describe{
    \item{getSamples}{\code{signature(obj = "mmctSamplerGeneric", ind = "numeric", n = "numeric")}: ... }
    \item{getNumber}{\code{signature(obj = "mmctSamplerGeneric")}: ... }
	 }
}
\references{
  Gandy, A. and Hahn, G. (2014) MMCTest - a safe algorithm for
  implementing multiple Monte Carlo tests. Scandinavian Journal of
  Statistics, 41(4):1083--1101
}
\author{ Axel Gandy and Georg Hahn }
\keyword{classes}
